Move schema v2 models into main package
