+++ src/retrocast/models/analysis.py
