--- src/retrocast/v2/models/__init__.py
@@ -1,62 +1,0 @@
-"""Schema v2 models."""
-
-from retrocast.v2.models.analysis import AnalysisReport, MetricSummary
-from retrocast.v2.models.candidates import Candidate, FailureRecord
-from retrocast.v2.models.evaluation import (
-    CheckResult,
-    CheckStatus,
-    ConstraintResult,
-    Evaluation,
-    ReactionValidity,
-    RouteValidity,
-    ScoredCandidate,
-    TargetResult,
-    Tier,
-    TierResult,
-)
-from retrocast.v2.models.route import (
-    InChIKeyLevel,
-    Molecule,
-    MoleculeId,
-    MoleculeView,
-    Reaction,
-    ReactionId,
-    ReactionView,
-    Route,
-    RoutePath,
-    validate_molecule_id,
-    validate_reaction_id,
-)
-from retrocast.v2.models.task import Benchmark, Target, Task, TaskConstraints
-
-__all__ = [
-    "AnalysisReport",
-    "Benchmark",
-    "Candidate",
-    "CheckResult",
-    "CheckStatus",
-    "ConstraintResult",
-    "Evaluation",
-    "FailureRecord",
-    "InChIKeyLevel",
-    "Molecule",
-    "MoleculeId",
-    "MoleculeView",
-    "MetricSummary",
-    "Reaction",
-    "ReactionId",
-    "ReactionValidity",
-    "ReactionView",
-    "Route",
-    "RoutePath",
-    "RouteValidity",
-    "ScoredCandidate",
-    "Target",
-    "TargetResult",
-    "Task",
-    "TaskConstraints",
-    "Tier",
-    "TierResult",
-    "validate_molecule_id",
-    "validate_reaction_id",
-]
