--- tests/models/test_route.py
@@ -1,610 +1,0 @@
-"""Unit tests for Route class."""
-
-import pytest
-
-from retrocast.models.chem import Molecule, PredictedRoute, ReactionStep, Route
-from retrocast.typing import InchiKeyStr, SmilesStr
-
-# ==============================================================================
-# Route Tests
-# ==============================================================================
-
-
-@pytest.mark.unit
-class TestRoute:
-    """Tests for the Route class."""
-
-    def test_basic_instantiation(self):
-        """Test creating a basic Route."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target)
-        assert route.target == target
-        assert route.metadata == {}
-
-    def test_depth_single_leaf(self):
-        """Test depth calculation for single leaf molecule."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target)
-        assert route.length == 0
-
-    def test_depth_single_step(self):
-        """Test depth calculation for single synthesis step."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant1, reactant2])
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
-        assert route.length == 1
-
-    def test_depth_multi_step_linear(self):
-        """Test depth calculation for multi-step linear route."""
-        # Build from bottom up
-        leaf = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-
-        intermediate1 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf]),
-        )
-
-        intermediate2 = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("FAKE-KEY-1"),
-            synthesis_step=ReactionStep(reactants=[intermediate1]),
-        )
-
-        target = Molecule(
-            smiles=SmilesStr("COCOC"),
-            inchikey=InchiKeyStr("FAKE-KEY-2"),
-            synthesis_step=ReactionStep(reactants=[intermediate2]),
-        )
-
-        route = Route(target=target)
-        assert route.length == 3
-
-    def test_iter_steps_and_reactions_are_root_first(self):
-        leaf = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        intermediate = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf], mapped_smiles="C>>CO"),
-        )
-        target = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("FAKE-KEY-1"),
-            synthesis_step=ReactionStep(reactants=[intermediate], mapped_smiles="CO>>COC"),
-        )
-        route = Route(target=target)
-
-        assert [step.mapped_smiles for step in route.iter_steps()] == ["CO>>COC", "C>>CO"]
-        assert [(reaction.product.smiles, reaction.step.mapped_smiles) for reaction in route.iter_reactions()] == [
-            ("COC", "CO>>COC"),
-            ("CO", "C>>CO"),
-        ]
-        assert [reaction.signature for reaction in route.iter_reactions()] == [
-            (frozenset(["OKKJLVBELUTLKV-UHFFFAOYSA-N"]), "FAKE-KEY-1"),
-            (frozenset(["VNWKTOKETHGBQD-UHFFFAOYSA-N"]), "OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-        ]
-
-    def test_reaction_ids_are_route_local_paths(self):
-        leaf_a = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("A"))
-        leaf_b = Molecule(smiles=SmilesStr("CC"), inchikey=InchiKeyStr("B"))
-        leaf_c = Molecule(smiles=SmilesStr("CCC"), inchikey=InchiKeyStr("C"))
-        left = Molecule(
-            smiles=SmilesStr("CCCC"),
-            inchikey=InchiKeyStr("LEFT"),
-            synthesis_step=ReactionStep(reactants=[leaf_a, leaf_b]),
-        )
-        right = Molecule(
-            smiles=SmilesStr("CCCCC"),
-            inchikey=InchiKeyStr("RIGHT"),
-            synthesis_step=ReactionStep(reactants=[leaf_c]),
-        )
-        target = Molecule(
-            smiles=SmilesStr("CCCCCC"),
-            inchikey=InchiKeyStr("TARGET"),
-            synthesis_step=ReactionStep(reactants=[left, right]),
-        )
-        route = Route(target=target)
-
-        reactions = list(route.iter_reactions())
-
-        assert [reaction.reaction_id for reaction in reactions] == ["rc:r:_", "rc:r:0", "rc:r:1"]
-        assert [reaction.product_id for reaction in reactions] == ["rc:m:_", "rc:m:0", "rc:m:1"]
-        assert route.get_reaction_by_id("rc:r:0").product.inchikey == "LEFT"
-
-    def test_depth_branched_route(self):
-        """Test depth calculation for branched route (should return max depth)."""
-        # Left branch: depth 2
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        intermediate_left = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-
-        # Right branch: depth 1 (just a leaf)
-        leaf2 = Molecule(smiles=SmilesStr("N"), inchikey=InchiKeyStr("QGZKDVFQNNGYKY-UHFFFAOYSA-N"))
-
-        # Combine branches
-        target = Molecule(
-            smiles=SmilesStr("CON"),
-            inchikey=InchiKeyStr("FAKE-KEY-3"),
-            synthesis_step=ReactionStep(reactants=[intermediate_left, leaf2]),
-        )
-
-        route = Route(target=target)
-        # Max depth: 1 (to intermediate_left) + 1 (to leaf1) = 2 for left branch
-        # Right branch is just 1 (to leaf2)
-        # So max is 2
-        assert route.length == 2
-
-    def test_leaves_property_single_leaf(self):
-        """Test leaves property for single leaf molecule."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target)
-        assert route.leaves == {target}
-        assert len(route.leaves) == 1
-
-    def test_leaves_property_simple_route(self):
-        """Test leaves property for simple route."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant1, reactant2])
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
-        assert route.leaves == {reactant1, reactant2}
-        assert len(route.leaves) == 2
-
-    def test_leaves_property_deduplication(self):
-        """Test that leaves property deduplicates correctly."""
-        # Same reactant used in different branches
-        common_reactant = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        unique_reactant = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-
-        step = ReactionStep(reactants=[common_reactant, common_reactant, unique_reactant])
-        target = Molecule(
-            smiles=SmilesStr("FAKE-PRODUCT"),
-            inchikey=InchiKeyStr("FAKE-KEY-4"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
-        assert len(route.leaves) == 2  # Deduplicated
-        assert common_reactant in route.leaves
-        assert unique_reactant in route.leaves
-
-    def test_get_signature_deterministic(self):
-        """Test that get_signature produces deterministic results."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant1, reactant2])
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
-
-        sig1 = route.get_structural_signature()
-        sig2 = route.get_structural_signature()
-        assert sig1 == sig2
-        assert isinstance(sig1, str)
-        assert len(sig1) == 64  # SHA256 hex digest length
-
-    def test_get_signature_identical_routes(self):
-        """Test that identical routes have same signature."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-
-        # Create first route
-        step1 = ReactionStep(reactants=[reactant1, reactant2])
-        target1 = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step1,
-        )
-        route1 = Route(target=target1)
-
-        # Create identical second route (different objects, same structure)
-        reactant1_copy = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2_copy = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step2 = ReactionStep(reactants=[reactant1_copy, reactant2_copy])
-        target2 = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step2,
-        )
-        route2 = Route(target=target2)
-
-        assert route1.get_structural_signature() == route2.get_structural_signature()
-
-    def test_get_signature_different_routes(self):
-        """Test that different routes have different signatures."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-
-        # First route
-        step1 = ReactionStep(reactants=[reactant1, reactant2])
-        target1 = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step1,
-        )
-        route1 = Route(target=target1)
-
-        # Different route (different reactants)
-        reactant3 = Molecule(
-            smiles=SmilesStr("C"),
-            inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"),
-        )
-        step2 = ReactionStep(reactants=[reactant3])
-        target2 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=step2,
-        )
-        route2 = Route(target=target2)
-
-        assert route1.get_structural_signature() != route2.get_structural_signature()
-
-    def test_get_signature_order_invariance(self):
-        """Test that reactant order doesn't affect signature."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-
-        # First order
-        step1 = ReactionStep(reactants=[reactant1, reactant2])
-        target1 = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step1,
-        )
-        route1 = Route(target=target1)
-
-        # Reversed order
-        reactant1_copy = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2_copy = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step2 = ReactionStep(reactants=[reactant2_copy, reactant1_copy])  # Reversed
-        target2 = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step2,
-        )
-        route2 = Route(target=target2)
-
-        # Should be the same due to sorting in get_signature
-        assert route1.get_structural_signature() == route2.get_structural_signature()
-
-    def test_get_signature_with_repeated_molecule(self):
-        """Test get_signature with the same molecule appearing multiple times (tests memoization)."""
-        # Branch 1: intermediate is formed from leaf1
-        leaf1 = Molecule(
-            smiles=SmilesStr("C"),
-            inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"),
-        )
-        branch1_intermediate = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-
-        # Branch 2: same molecule (by InChIKey) formed from leaf2
-        # This tests memoization in get_signature when the same InChIKey appears in different branches
-        leaf2 = Molecule(
-            smiles=SmilesStr("O"),
-            inchikey=InchiKeyStr("XLYOFNOQVPJJNP-UHFFFAOYSA-M"),
-        )
-        branch2_intermediate = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf2]),
-        )
-
-        # Combine branches into final product
-        target = Molecule(
-            smiles=SmilesStr("CCOCCO"),
-            inchikey=InchiKeyStr("MTHSVFCYNBDYFN-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[branch1_intermediate, branch2_intermediate]),
-        )
-
-        route = Route(target=target)
-        signature = route.get_structural_signature()
-
-        # Should produce a valid signature
-        assert isinstance(signature, str)
-        assert len(signature) == 64
-
-    def test_metadata_handling(self):
-        """Test route-level metadata."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        metadata = {"total_score": 0.95, "search_time": 42.5}
-        route = Route(target=target, metadata=metadata)
-        assert route.metadata["total_score"] == 0.95
-        assert route.metadata["search_time"] == 42.5
-
-    def test_has_convergent_reaction_single_leaf(self):
-        """Test that single leaf molecule has no convergent reaction."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target)
-        assert route.has_convergent_reaction is False
-
-
-@pytest.mark.unit
-class TestPredictedRoute:
-    """Tests for the PredictedRoute envelope."""
-
-    def test_from_route_preserves_zero_score_from_scores_metadata(self):
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target, metadata={"scores": {"state score": 0.0}})
-
-        prediction = PredictedRoute.from_route(route, rank=1)
-
-        assert prediction.rank == 1
-        assert prediction.score == 0.0
-        assert prediction.metadata == {"scores": {"state score": 0.0}}
-        assert prediction.route.metadata == {}
-
-    def test_from_route_strips_prediction_only_metadata_from_wrapped_route(self):
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(
-            target=target,
-            metadata={
-                "score": 0.9,
-                "confidence": 0.8,
-                "state_score": 0.7,
-                "source": "kept",
-            },
-        )
-
-        prediction = PredictedRoute.from_route(route)
-
-        assert prediction.score == 0.9
-        assert prediction.confidence == 0.8
-        assert prediction.metadata["state_score"] == 0.7
-        assert prediction.route.metadata == {"source": "kept"}
-        assert (
-            prediction.route.get_content_hash() == Route(target=target, metadata={"source": "kept"}).get_content_hash()
-        )
-
-    def test_content_hash_includes_prediction_envelope(self):
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target, metadata={"source": "kept"})
-
-        first = PredictedRoute.from_route(route, rank=1, metadata={"source_key": "target-1"})
-        second = PredictedRoute.from_route(route, rank=2, metadata={"source_key": "target-2"})
-
-        assert first.get_content_hash() != second.get_content_hash()
-        assert first.get_route_content_hash() == route.get_content_hash()
-        assert second.get_route_content_hash() == route.get_content_hash()
-
-    def test_from_route_uses_explicit_envelope_metadata_for_typed_prediction_fields(self):
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target, metadata={"source": "kept"})
-
-        prediction = PredictedRoute.from_route(
-            route,
-            metadata={"score": 0.9, "confidence": 0.8, "scores": {"state score": 0.7}},
-        )
-
-        assert prediction.score == 0.9
-        assert prediction.confidence == 0.8
-        assert prediction.metadata["score"] == 0.9
-        assert prediction.metadata["confidence"] == 0.8
-        assert prediction.route.metadata == {"source": "kept"}
-
-    def test_has_convergent_reaction_linear_route(self):
-        """Test that linear route (single reactant per step) has no convergent reaction."""
-        # Build linear route: A → B → C → D
-        leaf = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-
-        intermediate1 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf]),
-        )
-
-        intermediate2 = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("FAKE-KEY-1"),
-            synthesis_step=ReactionStep(reactants=[intermediate1]),
-        )
-
-        target = Molecule(
-            smiles=SmilesStr("COCOC"),
-            inchikey=InchiKeyStr("FAKE-KEY-2"),
-            synthesis_step=ReactionStep(reactants=[intermediate2]),
-        )
-
-        route = Route(target=target)
-        assert route.has_convergent_reaction is False
-
-    def test_has_convergent_reaction_two_leaves_combine(self):
-        """Test that combining two leaves is not convergent."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant1, reactant2])
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
-        assert route.has_convergent_reaction is False
-
-    def test_has_convergent_reaction_two_intermediates_combine(self):
-        """Test that combining two intermediates is convergent."""
-        # Branch 1: leaf1 → intermediate1
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        intermediate1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-
-        # Branch 2: leaf2 → intermediate2
-        leaf2 = Molecule(smiles=SmilesStr("O"), inchikey=InchiKeyStr("XLYOFNOQVPJJNP-UHFFFAOYSA-M"))
-        intermediate2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf2]),
-        )
-
-        # Convergent step: combine two intermediates
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[intermediate1, intermediate2]),
-        )
-
-        route = Route(target=target)
-        assert route.has_convergent_reaction is True
-
-    def test_has_convergent_reaction_nested_convergence(self):
-        """Test detection of convergent reaction deeper in the tree."""
-        # Build convergent step deep in tree:
-        # leaf1 → int1, leaf2 → int2, (int1 + int2) → int3, int3 → target
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        leaf2 = Molecule(smiles=SmilesStr("O"), inchikey=InchiKeyStr("XLYOFNOQVPJJNP-UHFFFAOYSA-M"))
-
-        intermediate1 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-        intermediate2 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf2]),
-        )
-
-        # Convergent step: int1 + int2 → int3
-        intermediate3 = Molecule(
-            smiles=SmilesStr("COCO"),
-            inchikey=InchiKeyStr("FAKE-CONV-KEY"),
-            synthesis_step=ReactionStep(reactants=[intermediate1, intermediate2]),
-        )
-
-        # Linear step: int3 → target
-        target = Molecule(
-            smiles=SmilesStr("COCOC"),
-            inchikey=InchiKeyStr("FAKE-TARGET-KEY"),
-            synthesis_step=ReactionStep(reactants=[intermediate3]),
-        )
-
-        route = Route(target=target)
-        assert route.has_convergent_reaction is True
-
-    def test_has_convergent_reaction_one_intermediate_one_leaf(self):
-        """Test that combining one intermediate and one leaf is not convergent."""
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        intermediate = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-
-        leaf2 = Molecule(smiles=SmilesStr("O"), inchikey=InchiKeyStr("XLYOFNOQVPJJNP-UHFFFAOYSA-M"))
-
-        target = Molecule(
-            smiles=SmilesStr("COO"),
-            inchikey=InchiKeyStr("FAKE-TARGET-KEY"),
-            synthesis_step=ReactionStep(reactants=[intermediate, leaf2]),
-        )
-
-        route = Route(target=target)
-        assert route.has_convergent_reaction is False
