+++ tests/models/test_route_signatures.py
@@ -1,285 +1,249 @@
-"""Unit tests for Route.get_reaction_signatures() method."""
-
-from typing import Any
-
 import pytest
+from hypothesis import given
+from hypothesis import strategies as st
+from pydantic import ValidationError
 
-from retrocast.models.chem import Molecule, ReactionSignature, ReactionStep, Route
-from retrocast.typing import InchiKeyStr, SmilesStr
+from retrocast.models.route import InChIKeyLevel, Molecule, Reaction, Route
+from retrocast.typing import InChIKeyStr, SmilesStr
 
-# ==============================================================================
-# Route.get_reaction_signatures Tests
-# ==============================================================================
+KEY_A = "AAAAAAAAAAAAAA-UHFFFAOYSA-N"
+KEY_B = "BBBBBBBBBBBBBB-UHFFFAOYSA-N"
+KEY_C = "CCCCCCCCCCCCCC-UHFFFAOYSA-N"
+KEY_D = "DDDDDDDDDDDDDD-UHFFFAOYSA-N"
+KEY_E = "EEEEEEEEEEEEEE-UHFFFAOYSA-N"
+KEY_A_STEREO_1 = "AAAAAAAAAAAAAA-BBBBBBBBSA-N"
+KEY_A_STEREO_2 = "AAAAAAAAAAAAAA-CCCCCCCCSA-N"
+
+
+def molecule(smiles: str, inchikey: str, product_of: Reaction | None = None) -> Molecule:
+    return Molecule(smiles=SmilesStr(smiles), inchikey=InChIKeyStr(inchikey), product_of=product_of)
+
+
+def one_step_route(reactants: list[Molecule], *, target_key: str = KEY_C) -> Route:
+    target = molecule("CC", target_key, product_of=Reaction(reactants=reactants))
+    return Route(target=target)
+
+
+def two_step_route() -> Route:
+    leaf = molecule("C", KEY_A)
+    intermediate = molecule("CO", KEY_B, product_of=Reaction(reactants=[leaf]))
+    target = molecule("CCO", KEY_C, product_of=Reaction(reactants=[intermediate, molecule("N", KEY_D)]))
+    return Route(target=target)
 
 
 @pytest.mark.unit
-class TestRouteGetReactionSignatures:
-    """Tests for Route.get_reaction_signatures() method."""
+def test_path_navigation_and_lookup_return_contextual_views() -> None:
+    route = two_step_route()
 
-    def test_leaf_route_returns_empty_set(self):
-        """Test that a route with no reactions returns empty set."""
-        target = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        route = Route(target=target)
+    target = route.molecule_at("rc:m:/")
+    root_reaction = route.reaction_at("rc:r:/")
+    intermediate = route.molecule_at("rc:m:/0")
+    child_reaction = route.reaction_at("rc:r:/0")
 
-        signatures = route.get_reaction_signatures()
-        assert signatures == set()
-        assert isinstance(signatures, set)
+    assert target.id() == "rc:m:/"
+    assert target.produced_by() == root_reaction
+    assert [reactant.id() for reactant in root_reaction.reactants()] == ["rc:m:/0", "rc:m:/1"]
+    assert root_reaction.id() == "rc:r:/"
+    assert intermediate.value.inchikey == KEY_B
+    assert child_reaction.product() == intermediate
+    assert child_reaction.reactants()[0].id() == "rc:m:/0/0"
 
-    def test_single_step_route(self):
-        """Test a route with single reaction step."""
-        reactant1 = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        reactant2 = Molecule(
-            smiles=SmilesStr("CC(=O)O"),
-            inchikey=InchiKeyStr("QTBSBXVTEAMEQO-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant1, reactant2])
-        target = Molecule(
-            smiles=SmilesStr("CCOC(C)=O"),
-            inchikey=InchiKeyStr("XEKOWRVHYACXOJ-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
 
-        signatures = route.get_reaction_signatures()
+@pytest.mark.unit
+def test_missing_molecule_or_reaction_lookup_raises_key_error() -> None:
+    route = one_step_route([molecule("C", KEY_A)])
 
-        assert len(signatures) == 1
-        sig = next(iter(signatures))
-        assert isinstance(sig, tuple)
-        assert len(sig) == 2
-        reactant_keys, product_key = sig
-        assert isinstance(reactant_keys, frozenset)
-        assert product_key == "XEKOWRVHYACXOJ-UHFFFAOYSA-N"
-        assert reactant_keys == frozenset(["LFQSCWFLJHTTHZ-UHFFFAOYSA-N", "QTBSBXVTEAMEQO-UHFFFAOYSA-N"])
+    with pytest.raises(ValueError):
+        route.molecule_at("rc:r:/")
+    with pytest.raises(ValueError):
+        route.reaction_at("rc:m:/")
+    with pytest.raises(KeyError):
+        route.molecule_at("rc:m:/9")
+    with pytest.raises(KeyError) as exc_info:
+        route.reaction_at("rc:r:/0")
+    assert exc_info.value.args == ("rc:r:/0",)
 
-    def test_multi_step_linear_route(self):
-        """Test a linear route with multiple steps."""
-        # Build from bottom up: leaf -> intermediate -> target
-        leaf = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
 
-        intermediate = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf]),
-        )
+@pytest.mark.unit
+def test_lookup_past_leaf_raises_key_error() -> None:
+    route = one_step_route([molecule("C", KEY_A)])
 
-        target = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("FAKE-KEY-1"),
-            synthesis_step=ReactionStep(reactants=[intermediate]),
-        )
+    with pytest.raises(KeyError):
+        route.molecule_at("rc:m:/0/0")
 
-        route = Route(target=target)
-        signatures = route.get_reaction_signatures()
 
-        assert len(signatures) == 2
+@pytest.mark.unit
+def test_reaction_lookup_past_leaf_raises_requested_reaction_id() -> None:
+    route = one_step_route([molecule("C", KEY_A)])
 
-        # Check both reactions are present
-        expected_sig1: ReactionSignature = (frozenset(["VNWKTOKETHGBQD-UHFFFAOYSA-N"]), "OKKJLVBELUTLKV-UHFFFAOYSA-N")
-        expected_sig2: ReactionSignature = (frozenset(["OKKJLVBELUTLKV-UHFFFAOYSA-N"]), "FAKE-KEY-1")
+    with pytest.raises(KeyError) as exc_info:
+        route.reaction_at("rc:r:/0/0")
+    assert exc_info.value.args == ("rc:r:/0/0",)
 
-        assert expected_sig1 in signatures
-        assert expected_sig2 in signatures
 
-    def test_branched_route(self):
-        """Test a branched route with multiple parallel reactions."""
-        # Left branch
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"))
-        intermediate_left = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("OKKJLVBELUTLKV-UHFFFAOYSA-N"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
+@pytest.mark.unit
+def test_leaf_route_has_target_but_no_root_reaction() -> None:
+    route = Route(target=molecule("C", KEY_A))
 
-        # Right branch (just a leaf)
-        leaf2 = Molecule(smiles=SmilesStr("N"), inchikey=InchiKeyStr("QGZKDVFQNNGYKY-UHFFFAOYSA-N"))
+    assert route.molecule_at("rc:m:/").value.inchikey == KEY_A
+    assert route.molecule_at("rc:m:/").produced_by() is None
+    assert [leaf.id() for leaf in route.leaves()] == ["rc:m:/"]
+    assert [leaf.id() for leaf in route.iter_leaves()] == ["rc:m:/"]
+    assert route.depth() == 0
+    with pytest.raises(KeyError):
+        route.reaction_at("rc:r:/")
 
-        # Combine branches
-        target = Molecule(
-            smiles=SmilesStr("CON"),
-            inchikey=InchiKeyStr("FAKE-KEY-3"),
-            synthesis_step=ReactionStep(reactants=[intermediate_left, leaf2]),
-        )
 
-        route = Route(target=target)
-        signatures = route.get_reaction_signatures()
+@pytest.mark.unit
+def test_route_leaves_and_depth_use_route_paths() -> None:
+    route = two_step_route()
 
-        # Should have 2 reactions: leaf1->intermediate_left, (intermediate_left + leaf2)->target
-        assert len(signatures) == 2
+    assert [leaf.id() for leaf in route.leaves()] == ["rc:m:/0/0", "rc:m:/1"]
+    assert [leaf.id() for leaf in route.iter_leaves()] == ["rc:m:/0/0", "rc:m:/1"]
+    assert route.depth() == 2
 
-        expected_sig1: ReactionSignature = (frozenset(["VNWKTOKETHGBQD-UHFFFAOYSA-N"]), "OKKJLVBELUTLKV-UHFFFAOYSA-N")
-        expected_sig2: ReactionSignature = (
-            frozenset(["OKKJLVBELUTLKV-UHFFFAOYSA-N", "QGZKDVFQNNGYKY-UHFFFAOYSA-N"]),
-            "FAKE-KEY-3",
-        )
 
-        assert expected_sig1 in signatures
-        assert expected_sig2 in signatures
+@pytest.mark.unit
+def test_reaction_identity_includes_product_and_reactants() -> None:
+    route_a = one_step_route([molecule("C", KEY_A), molecule("O", KEY_B)], target_key=KEY_C)
+    route_b = one_step_route([molecule("C", KEY_A), molecule("O", KEY_B)], target_key=KEY_D)
+    route_c = one_step_route([molecule("C", KEY_A), molecule("N", KEY_E)], target_key=KEY_C)
 
-    def test_reaction_signature_is_hashable(self):
-        """Test that reaction signatures can be added to sets and used as dict keys."""
-        sig: ReactionSignature = (frozenset(["KEY1", "KEY2"]), "KEY3")
+    assert route_a.reaction_at("rc:r:/").signature() != route_b.reaction_at("rc:r:/").signature()
+    assert route_a.reaction_at("rc:r:/").signature() != route_c.reaction_at("rc:r:/").signature()
 
-        # Can be added to set
-        sig_set = {sig}
-        assert sig in sig_set
 
-        # Can be used as dict key
-        sig_dict = {sig: "test_value"}
-        assert sig_dict[sig] == "test_value"
+@pytest.mark.unit
+def test_reactant_order_does_not_affect_route_signature() -> None:
+    route_a = one_step_route([molecule("C", KEY_A), molecule("O", KEY_B)])
+    route_b = one_step_route([molecule("O", KEY_B), molecule("C", KEY_A)])
 
-    def test_duplicate_reaction_in_route(self):
-        """Test that same reaction appearing multiple times is deduplicated."""
-        # This is an edge case where the same molecule (by InchiKey) is used in multiple places
-        # but each has different synthesis paths - the reactions should be collected
+    assert route_a.signature() == route_b.signature()
 
-        # Create a diamond-shaped dependency
-        leaf1 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("LEAF1-KEY"))
-        leaf2 = Molecule(smiles=SmilesStr("O"), inchikey=InchiKeyStr("LEAF2-KEY"))
 
-        # Both branches produce intermediates
-        intermediate1 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("INTERMEDIATE1-KEY"),
-            synthesis_step=ReactionStep(reactants=[leaf1]),
-        )
-        intermediate2 = Molecule(
-            smiles=SmilesStr("OC"),
-            inchikey=InchiKeyStr("INTERMEDIATE2-KEY"),
-            synthesis_step=ReactionStep(reactants=[leaf2]),
-        )
+@pytest.mark.unit
+def test_reactant_order_does_not_affect_reaction_signature() -> None:
+    route_a = one_step_route([molecule("C", KEY_A), molecule("O", KEY_B)])
+    route_b = one_step_route([molecule("O", KEY_B), molecule("C", KEY_A)])
 
-        # Final product combines both intermediates
-        target = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("TARGET-KEY"),
-            synthesis_step=ReactionStep(reactants=[intermediate1, intermediate2]),
-        )
+    assert route_a.reaction_at("rc:r:/").signature() == route_b.reaction_at("rc:r:/").signature()
 
-        route = Route(target=target)
-        signatures = route.get_reaction_signatures()
 
-        # Should have 3 unique reactions
-        assert len(signatures) == 3
+@pytest.mark.unit
+def test_duplicate_reactants_do_affect_route_signature() -> None:
+    route_a = one_step_route([molecule("C", KEY_A)])
+    route_b = one_step_route([molecule("C", KEY_A), molecule("C", KEY_A)])
 
-    def test_comparing_routes_for_overlap(self):
-        """Test the main use case: finding if routes share any reactions."""
-        # Route 1: A -> B -> C
-        leaf_a = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("KEY-A"))
-        intermediate_b = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("KEY-B"),
-            synthesis_step=ReactionStep(reactants=[leaf_a]),
-        )
-        target_c = Molecule(
-            smiles=SmilesStr("COC"),
-            inchikey=InchiKeyStr("KEY-C"),
-            synthesis_step=ReactionStep(reactants=[intermediate_b]),
-        )
-        route1 = Route(target=target_c)
+    assert route_a.signature() != route_b.signature()
 
-        # Route 2: A -> B -> D (shares first reaction with route1)
-        leaf_a2 = Molecule(smiles=SmilesStr("C"), inchikey=InchiKeyStr("KEY-A"))
-        intermediate_b2 = Molecule(
-            smiles=SmilesStr("CO"),
-            inchikey=InchiKeyStr("KEY-B"),
-            synthesis_step=ReactionStep(reactants=[leaf_a2]),
-        )
-        target_d = Molecule(
-            smiles=SmilesStr("COCC"),
-            inchikey=InchiKeyStr("KEY-D"),
-            synthesis_step=ReactionStep(reactants=[intermediate_b2]),
-        )
-        route2 = Route(target=target_d)
 
-        # Route 3: X -> Y -> Z (no overlap with route1 or route2)
-        leaf_x = Molecule(smiles=SmilesStr("N"), inchikey=InchiKeyStr("KEY-X"))
-        intermediate_y = Molecule(
-            smiles=SmilesStr("NC"),
-            inchikey=InchiKeyStr("KEY-Y"),
-            synthesis_step=ReactionStep(reactants=[leaf_x]),
-        )
-        target_z = Molecule(
-            smiles=SmilesStr("NCC"),
-            inchikey=InchiKeyStr("KEY-Z"),
-            synthesis_step=ReactionStep(reactants=[intermediate_y]),
-        )
-        route3 = Route(target=target_z)
+@pytest.mark.unit
+def test_duplicate_reactants_do_affect_reaction_signature() -> None:
+    route_a = one_step_route([molecule("C", KEY_A)])
+    route_b = one_step_route([molecule("C", KEY_A), molecule("C", KEY_A)])
 
-        sigs1 = route1.get_reaction_signatures()
-        sigs2 = route2.get_reaction_signatures()
-        sigs3 = route3.get_reaction_signatures()
+    assert route_a.reaction_at("rc:r:/").signature() != route_b.reaction_at("rc:r:/").signature()
 
-        # Routes 1 and 2 share the A->B reaction
-        overlap_1_2 = sigs1 & sigs2
-        assert len(overlap_1_2) == 1
-        shared_reaction = next(iter(overlap_1_2))
-        assert shared_reaction == (frozenset(["KEY-A"]), "KEY-B")
 
-        # Routes 1 and 3 have no overlap
-        overlap_1_3 = sigs1 & sigs3
-        assert len(overlap_1_3) == 0
+@pytest.mark.unit
+def test_prefix_depth_changes_signatures_as_expected() -> None:
+    shallow_intermediate = molecule("CO", KEY_B)
+    deep_intermediate = molecule("CO", KEY_B, product_of=Reaction(reactants=[molecule("C", KEY_A)]))
+    shallow_route = one_step_route([shallow_intermediate, molecule("N", KEY_D)])
+    deep_route = one_step_route([deep_intermediate, molecule("N", KEY_D)])
 
-        # Routes 2 and 3 have no overlap
-        overlap_2_3 = sigs2 & sigs3
-        assert len(overlap_2_3) == 0
+    assert shallow_route.signature(depth=0) == deep_route.signature(depth=0)
+    assert shallow_route.signature(depth=1) == deep_route.signature(depth=1)
+    assert shallow_route.signature() != deep_route.signature()
 
-    def test_reaction_signature_type_alias(self):
-        """Test that ReactionSignature type alias works correctly."""
-        # Create a valid ReactionSignature
-        sig: ReactionSignature = (frozenset(["KEY1", "KEY2"]), "KEY3")
 
-        # Verify structure
-        reactants, product = sig
-        assert isinstance(reactants, frozenset)
-        assert isinstance(product, str)
+@pytest.mark.unit
+def test_subtree_signature_addresses_non_root_molecules() -> None:
+    route = two_step_route()
 
-    def test_deep_route_with_pharma_data(self, pharma_routes_data: dict[str, Any]):
-        """Test get_reaction_signatures with real pharma route data."""
-        from tests.models.test_pharma_routes import TestPharmaRoutesContract
+    assert route.molecule_at("rc:m:/0").subtree_signature() != route.molecule_at("rc:m:/1").subtree_signature()
+    assert route.molecule_at("rc:m:/0").value.signature() == molecule("CO", KEY_B).signature()
+    assert route.molecule_at("rc:m:/0").subtree_key(depth=0) == ("mol", KEY_B)
 
-        helper = TestPharmaRoutesContract()
-        vonoprazan_data = pharma_routes_data["vonoprazan-1"]
-        target_molecule = helper._build_molecule_tree(vonoprazan_data)
-        route = Route(target=target_molecule)
 
-        signatures = route.get_reaction_signatures()
+@pytest.mark.unit
+def test_equal_subtrees_in_different_positions_have_equal_signatures() -> None:
+    left = molecule("CO", KEY_B, product_of=Reaction(reactants=[molecule("C", KEY_A)]))
+    right = molecule("CO", KEY_B, product_of=Reaction(reactants=[molecule("C", KEY_A)]))
+    route = one_step_route([left, right])
 
-        # Vonoprazan has depth 2, so should have at least 2 reactions
-        assert len(signatures) >= 2
+    assert route.molecule_at("rc:m:/0").id() != route.molecule_at("rc:m:/1").id()
+    assert route.molecule_at("rc:m:/0").subtree_signature() == route.molecule_at("rc:m:/1").subtree_signature()
 
-        # All signatures should be valid tuples
-        for sig in signatures:
-            assert isinstance(sig, tuple)
-            assert len(sig) == 2
-            reactants, product = sig
-            assert isinstance(reactants, frozenset)
-            assert isinstance(product, str)
-            assert len(reactants) > 0  # Must have at least one reactant
 
-    def test_single_reactant_reaction(self):
-        """Test a reaction with a single reactant."""
-        reactant = Molecule(
-            smiles=SmilesStr("CCO"),
-            inchikey=InchiKeyStr("LFQSCWFLJHTTHZ-UHFFFAOYSA-N"),
-        )
-        step = ReactionStep(reactants=[reactant])
-        target = Molecule(
-            smiles=SmilesStr("C"),
-            inchikey=InchiKeyStr("VNWKTOKETHGBQD-UHFFFAOYSA-N"),
-            synthesis_step=step,
-        )
-        route = Route(target=target)
+@pytest.mark.unit
+def test_route_serialization_does_not_include_derived_node_ids() -> None:
+    route = two_step_route()
 
-        signatures = route.get_reaction_signatures()
+    payload = route.model_dump_json()
+    loaded = Route.model_validate_json(payload)
 
-        assert len(signatures) == 1
-        sig = next(iter(signatures))
-        reactant_keys, product_key = sig
-        assert reactant_keys == frozenset(["LFQSCWFLJHTTHZ-UHFFFAOYSA-N"])
-        assert product_key == "VNWKTOKETHGBQD-UHFFFAOYSA-N"
+    assert "rc:m:" not in payload
+    assert "rc:r:" not in payload
+    assert loaded.signature() == route.signature()
+
+
+@pytest.mark.unit
+def test_route_schema_version_is_literal_v2() -> None:
+    route = two_step_route()
+
+    assert route.schema_version == "2"
+    with pytest.raises(ValidationError):
+        Route.model_validate({"target": route.target.model_dump(mode="json"), "schema_version": "3"})
+
+
+@pytest.mark.unit
+def test_match_level_controls_molecule_identity() -> None:
+    route_a = Route(target=molecule("C", KEY_A_STEREO_1))
+    route_b = Route(target=molecule("C", KEY_A_STEREO_2))
+
+    assert route_a.signature(InChIKeyLevel.FULL) != route_b.signature(InChIKeyLevel.FULL)
+    assert route_a.signature(InChIKeyLevel.NO_STEREO) == route_b.signature(InChIKeyLevel.NO_STEREO)
+    assert route_a.signature(InChIKeyLevel.CONNECTIVITY) == route_b.signature(InChIKeyLevel.CONNECTIVITY)
+
+
+@pytest.mark.unit
+def test_negative_depth_is_rejected() -> None:
+    route = two_step_route()
+
+    with pytest.raises(ValueError):
+        route.key(depth=-1)
+    with pytest.raises(ValueError):
+        route.signature(depth=-1)
+    with pytest.raises(ValueError):
+        route.molecule_at("rc:m:/0").subtree_key(depth=-1)
+    with pytest.raises(ValueError):
+        route.molecule_at("rc:m:/0").subtree_signature(depth=-1)
+
+
+@pytest.mark.unit
+@given(st.permutations([KEY_A, KEY_B, KEY_D]))
+def test_generated_reactant_permutations_do_not_change_signature(reactant_keys: tuple[str, ...]) -> None:
+    reference = one_step_route([molecule("C", KEY_A), molecule("O", KEY_B), molecule("N", KEY_D)])
+    permuted = one_step_route([molecule("C", key) for key in reactant_keys])
+
+    assert permuted.signature() == reference.signature()
+
+
+@pytest.mark.unit
+@given(st.lists(st.sampled_from([KEY_A, KEY_B, KEY_D]), min_size=1, max_size=5))
+def test_generated_routes_keep_signature_across_serialization(reactant_keys: list[str]) -> None:
+    route = one_step_route([molecule("C", key) for key in reactant_keys])
+
+    loaded = Route.model_validate(route.model_dump(mode="json"))
+
+    assert loaded.signature() == route.signature()
+
+
+@pytest.mark.unit
+@given(st.lists(st.sampled_from([KEY_A, KEY_B, KEY_D]), min_size=1, max_size=5))
+def test_generated_duplicate_multiplicity_changes_signature(reactant_keys: list[str]) -> None:
+    route = one_step_route([molecule("C", key) for key in reactant_keys])
+    with_extra_duplicate = one_step_route([molecule("C", key) for key in [*reactant_keys, reactant_keys[0]]])
+
+    assert with_extra_duplicate.signature() != route.signature()
